backend/requirements.txt
fastapi[standard]
paho-mqtt
psycopg2-binary
pandas